database: subsong id with 16bit index playlist: fix load of playlist with deprecated subsong id

+++ source/Database/Types/SubsongID.inl
@@ -30,7 +30,7 @@ namespace rePlayer
         return Value() < other.Value();
     }
 
-    inline constexpr SubsongID::SubsongID(SongID newSongId, uint32_t newIndex)
+    inline constexpr SubsongID::SubsongID(SongID newSongId, uint16_t newIndex)
         : songId(newSongId)
         , index(newIndex)
     {}
